Renamed runtime weight params file to avoid confusion with generated weights files
